requirements.txt
Django>=3.2,<6.0
markdown>=3.4.0
Pygments>=2.12.0
whitenoise>=6.0.0
gunicorn>=20.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
django-mptt>=0.14.0
mysqlclient>=2.2.0
drf-yasg>=1.21.7
PyJWT>=2.8.0
